Supprime le lien d'accès panier actions impact de l'accès rapide du Header
