+++ app.territoiresentransitions.react/src/app/Layout/Header/useNbActionsDansPanier.ts
@@ -1,0 +1,16 @@
+import { useQuery } from 'react-query';
+
+import { PanierAPI } from '@/api';
+import { supabaseClient } from '@/app/core-logic/api/supabase';
+
+const panierAPI = new PanierAPI(supabaseClient);
+
+/**
+ * Charge le nombre d'actions présentes dans le panier d'une collectivité
+ */
+export const useNbActionsDansPanier = (collectiviteId: number | null) => {
+  return useQuery(['nb_actions_dans_panier', collectiviteId], async () => {
+    if (!collectiviteId) return;
+    return panierAPI.getCollectivitePanierInfo(collectiviteId);
+  });
+};
